requirements.txt
asgiref==3.2.10
Django==3.0.7
django-rest-framework==0.1.0
djangorestframework==3.11.0
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
zope.event==4.4
zope.interface==5.1.0
